Update location schema tests
